move server-only types to their own module and re-export into crate::server
