+++ rodbus/src/server/handler.rs
@@ -3,6 +3,7 @@ use std::sync::{Arc, Mutex};
 
 use crate::exception::ExceptionCode;
 use crate::types::*;
+use crate::server::{WriteCoils, WriteRegisters};
 
 /// Trait implemented by the user to process requests received from the client
 ///
