+++ rodbus/src/server/mod.rs
@@ -5,15 +5,19 @@ use tracing::Instrument;
 use crate::decode::DecodeLevel;
 use crate::tokio;
 
-use crate::server::handler::{RequestHandler, ServerHandlerMap};
 use crate::shutdown::TaskHandle;
 use crate::tcp::server::ServerTask;
 
 /// server handling
-pub mod handler;
+pub(crate) mod handler;
 pub(crate) mod request;
 pub(crate) mod response;
 pub(crate) mod task;
+pub(crate) mod types;
+
+// re-export to the public API
+pub use handler::*;
+pub use types::*;
 
 /// Spawns a TCP server task onto the runtime. This method can only
 /// be called from within the runtime context. Use [`create_tcp_server_task`]
