+++ rodbus/src/server/request.rs
@@ -5,6 +5,7 @@ use crate::common::traits::{Loggable, Parse, Serialize};
 use crate::decode::PduDecodeLevel;
 use crate::exception::ExceptionCode;
 use crate::error::Error;
+use crate::server::*;
 use crate::server::handler::RequestHandler;
 use crate::server::response::{BitWriter, RegisterWriter};
 use crate::types::*;
